+++ Changelog.md
@@ -4,7 +4,15 @@
 
 ## CHANGELOG
 
-### 0.4.0 @ 01/23/2026
+### 0.4.0 @ 01/26/2026
+
+#### :mega: New
+
+1. Add the extended image composers `stet.improc.composer`. This is the extended `ImageChops` module.
+2. Add the image effect `stet.improc.effects` which supports the blending effects of images.
+3. Add the image overlay effect `stet.improc.overlays` which supports blending effects covering the image content.
+4. Split `stet.improc.rederer.ImageLayer` and its related functionalities into `stet.improc.layer`, where the layer properties are implemented.
+5. Refactor the shadow/glow/stroke effects of the text images. Now, these effects are delegated to `ImageLayer`.
 
 #### :wrench: Fix